requirements.txt
flask==1.1.2
Flask-RESTful==0.3.8
Werkzeug==1.0.1
requests==2.25.1
PyYAML==5.4.1
APScheduler==3.7.0
